requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.2
django-allauth==0.36.0
django-model-utils==3.1.2
django-recaptcha==1.4.0
idna==2.8
mysqlclient==1.3.13
oauthlib==3.1.0
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
